requirements.txt
numpy
matplotlib
pillow
imageio
imageio-ffmpeg